feat: add @MemphisConsume() decorator
this decorator is kind of better than @ConsumeMessage().
both decorators do the same thing (receiving messages).
the difference is in the fact that this @MemphisConsume()
can work outside controller files (that's inside service files),
of which the other @ConsumeMessage() decorator can't.

@ConsumeMessage relies on a Pattern decorator from NestJS.
NestJS's pattern decorators (@MessagePattern() and @EventPattern())
work only in controllers (reason for the need of this @MemphisConsume())

also update README and example file to use @MemphisConsume() instead of @ConsumeMessage()